+++ kaithem/src/tests/test_events.py
@@ -9,7 +9,7 @@ def test_events():
     # Create an event that sets y to 0 if it is 1
     with newevt._event_list_lock:
         x = newevt.Event("y==1", "global y\ny=0", setup="y=0")
-        x.module = x.resource = "testevt"
+        x.module = x.resource = "pytest_testevt"
         newevt.EventReferences[x.module, x.resource] = x
 
         # Register event with polling.